requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
sqlalchemy==2.0.35
python-dotenv==1.0.1
aiofiles==24.1.0
